+++ cuda/inc/GeneralOperators/phasefilters.h
@@ -37,6 +37,8 @@ extern "C"{
 	void _rytov_gpu(PAR param, float *projections, float *d_kernel, size_t nrays, size_t nangles, size_t nslices);
 	void _bronnikov_gpu(PAR param, float *projections, float *d_kernel, size_t nrays, size_t nangles, size_t nslices);
 
+	__global__ void KCopy(cufftComplex *in, float *out, size_t sizex, size_t sizey, size_t sizez);
+
 }
 
 #endif